we all float down here

+++ api/src/schema/index.js
@@ -126,8 +126,8 @@ const Schema = i18n => {
     }
 
     type ErsRating @cacheControl(maxAge: 90) {
-      measurement: Int
-      upgrade: Int
+      measurement: Float
+      upgrade: Float
     }
 
     # ${i18n.t`The root query type`}
